ADD: update method to del item

+++ src/routes/stock.js
@@ -37,8 +37,14 @@ router.post('/api/stock', async (req, res) => {
 
 // DELETING ONE
 router.delete('/api/stock/:id', async (req, res) => {
+    const foundItem = await Stock.findOne({ _id: req.params.id })
+
+    if (!foundItem) {
+        console.log('Item NOT found!')
+        return
+    }
+
     try {
-        const foundItem = await Stock.findOne({ _id: req.params.id })
         await foundItem.delete()
         res.status(200).json({ message: 'Item deleted.' })
     } catch (err) {
